Updating crux from Neb dev.

+++ code/modules/mob/living/simple_animal/hostile/viscerator.dm
@@ -3,7 +3,7 @@
 	desc = "A small, twin-bladed machine capable of inflicting very deadly lacerations."
 	icon = 'icons/mob/simple_animal/viscerator.dmi'
 	pass_flags = PASS_FLAG_TABLE
-	mob_default_max_health = 15
+	max_health = 15
 	natural_weapon = /obj/item/natural_weapon/rotating_blade
 	faction = "syndicate"
 	min_gas = null
@@ -11,13 +11,7 @@
 	minbodytemp = 0
 	attack_delay = DEFAULT_QUICK_COOLDOWN
 	bleed_colour = SYNTH_BLOOD_COLOR
-
-	meat_type =     null
-	meat_amount =   0
-	bone_material = null
-	bone_amount =   0
-	skin_material = null
-	skin_amount =   0
+	butchery_data = /decl/butchery_data/synthetic
 
 /obj/item/natural_weapon/rotating_blade
 	name = "rotating blades"
@@ -30,9 +24,13 @@
 /mob/living/simple_animal/hostile/viscerator/check_has_mouth()
 	return FALSE
 
-/mob/living/simple_animal/hostile/viscerator/death(gibbed, deathmessage, show_dead_message)
-	..(null, "is smashed into pieces!", show_dead_message)
-	qdel(src)
+/mob/living/simple_animal/hostile/viscerator/get_death_message(gibbed)
+	return "is smashed into pieces!"
+
+/mob/living/simple_animal/hostile/viscerator/death(gibbed)
+	. = ..()
+	if(. && !gibbed)
+		qdel(src)
 
 /mob/living/simple_animal/hostile/viscerator/hive
 	faction = "hivebot"
